Fix testing infrastructure
There are two types of tests: ones meant to run directly on a CheriBSD
machine, and ones meant to be ran as a compartment, using a binary with
our library. Before, we were not actually running the compartment tests
as compartments, just doing so directly. This fixes this, as well as
slightly touching on the testing implementation design (no functional
changes).

Additionally fix some bugs with the code
* Fix not saving `x/c28` when coming from an interecept. Was messing
  with a stack variable when intercepting a `realloc` call.
* Add intercept for `fputc`. WHERE DOES IT COME FROM??? `lua_writeline`
  presumable resolves to `fwrite` (`lauxlib.c:258-266`), but the
  assembly branches to `fputc`?