Compartment configuration files
Now can specify compartment entry points via a `toml` file of the name
`<binary>.comp`, where `<binary>` is the name of the ELF binary to be
compartmentalize, and the associated configuration file should reside in
the same folder as the compartment binary.

The format of the `toml` file is:
* each entry point is defined as a `toml` table, with the table name
  being the entry function name in the compartment binary;
* entry `args_count` is of type number, and says how many arguments we
  expect to pass to the entry function;
* entry `args_type` is an array and enumerates the types to be passed to
  the function.

This comes with additional internal functionality to convert between
types (the conversion is done manually, i.e., new types need to be added
to the code, rather than being dynamically handled).

Additional tests added, and reworked how tests in CMake are done, to
better handle argument support.