Memory layout improvements
Move around compartment scratch memory, to ensure the heap is next to
the DDC boundary in the compartment. Thus, any overflows should trigger
a `SIGPROT`.

Ensure compartment boundaries are as they are defined in the struct.
Particularly, in `mmap`, `length` would be padded such that it would be
divisible by `page_size`. This padding is now done explicitly, and
additional checks are added throughout to ensure this remains
consistent. This prevents the compartment scratch heap to be larger than
configured in the compartment.

Further changes:
* `run_test.py` now always returns, without giving an exception
* added `simple_malloc_saturate`, which allocates all available heap
  memory, and then some, and expects to fail
* `malloc` in `comp_utils` now returns `NULL` when running out of
  memory, instead of calling `err`
* mark some code in `simple_call_internal_weak` to not be
  `clang-format`ted, as it seems to diverge from the CHERI version of
  `clang-format`