1 compartment to N mappings
Move from a design of "one compartment has one mapping" to "one
compartment can be mapped multiple times". This involves "staging" a
compartment, with temporary internal relative relocations, after which
the staging can be lifted to a "mapping", which has correct relocations
and can be executed. This greatly reduces the overhead of creating the
same compartment multiple times.

* Remove some now unused functions