Compartment reuse
Implement a "pseudo" compartment caching feature to evaluate performance
in a more ideal implementation. Rather than fully cleaning a mapping for
a compartment, we emulate keeping around the last copy, which precludes
a pair of `mmap` / `munmap`, as well as downsizing the amount of cleanup
needed to be done (we only need to reset the stack, the allocated parts
of the heap, and the writeable segments). Note that this is not a full
implementation, and it likely breaks with multiple compartments as it
currently is.

Major changes
* Move `mappings` code in a separate file
* Implement mapping reuse (includes caching a mapping, cleaning the
  stack and heap allocations, and resetting writeable segments)
* Add a benchmarking implementation; this is done as it is useful to
  have, and calls to the benchmarking API are not present in the tree,
  but can be added by users
* Allow only an entry point be specified in `manager_arg_passer`, rather
  than requiring arguments to be passed to that entry point
* Add some additional tests, including more failing ones (comment out
  `lua_suite_some`, as it crashes on the CI system)
* Other fixes